@@ -3,7 +3,7 @@
     <Description>A .NET wrapper for the ImGuizmo library.</Description>
     <AssemblyVersion>1.61.0</AssemblyVersion>
     <Authors>Eric Mellino</Authors>
-    <TargetFrameworks>netstandard2.0;net6.0</TargetFrameworks>
+    <TargetFrameworks>net7.0</TargetFrameworks>
     <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
     <DebugType>portable</DebugType>
     <AssemblyName>ImGuizmo.NET</AssemblyName>
