@@ -265,7 +265,7 @@
         /// <inheritdoc/>
         public override int GetHashCode()
         {
-            HashCode hashCode = new HashCode();
+            HashCode hashCode = new();
             {
                 hashCode.Add(Center);
                 hashCode.Add(Radius);
