@@ -2,10 +2,10 @@
 {
     public enum MapMode : int
     {
-        Read = unchecked((int)1),
-        Write = unchecked((int)2),
-        ReadWrite = unchecked((int)3),
-        WriteDiscard = unchecked((int)4),
-        WriteNoOverwrite = unchecked((int)5)
+        Read = unchecked(1),
+        Write = unchecked(2),
+        ReadWrite = unchecked(3),
+        WriteDiscard = unchecked(4),
+        WriteNoOverwrite = unchecked(5)
     }
 }
