@@ -2,47 +2,47 @@
 {
     public enum PrimitiveTopology : int
     {
-        Undefined = unchecked((int)0),
-        PointList = unchecked((int)1),
-        LineList = unchecked((int)2),
-        LineStrip = unchecked((int)3),
-        TriangleList = unchecked((int)4),
-        TriangleStrip = unchecked((int)5),
-        LineListAdjacency = unchecked((int)10),
-        LineStripAdjacency = unchecked((int)11),
-        TriangleListAdjacency = unchecked((int)12),
-        TriangleStripAdjacency = unchecked((int)13),
-        PatchListWith1ControlPoints = unchecked((int)33),
-        PatchListWith2ControlPoints = unchecked((int)34),
-        PatchListWith3ControlPoints = unchecked((int)35),
-        PatchListWith4ControlPoints = unchecked((int)36),
-        PatchListWith5ControlPoints = unchecked((int)37),
-        PatchListWith6ControlPoints = unchecked((int)38),
-        PatchListWith7ControlPoints = unchecked((int)39),
-        PatchListWith8ControlPoints = unchecked((int)40),
-        PatchListWith9ControlPoints = unchecked((int)41),
-        PatchListWith10ControlPoints = unchecked((int)42),
-        PatchListWith11ControlPoints = unchecked((int)43),
-        PatchListWith12ControlPoints = unchecked((int)44),
-        PatchListWith13ControlPoints = unchecked((int)45),
-        PatchListWith14ControlPoints = unchecked((int)46),
-        PatchListWith15ControlPoints = unchecked((int)47),
-        PatchListWith16ControlPoints = unchecked((int)48),
-        PatchListWith17ControlPoints = unchecked((int)49),
-        PatchListWith18ControlPoints = unchecked((int)50),
-        PatchListWith19ControlPoints = unchecked((int)51),
-        PatchListWith20ControlPoints = unchecked((int)52),
-        PatchListWith21ControlPoints = unchecked((int)53),
-        PatchListWith22ControlPoints = unchecked((int)54),
-        PatchListWith23ControlPoints = unchecked((int)55),
-        PatchListWith24ControlPoints = unchecked((int)56),
-        PatchListWith25ControlPoints = unchecked((int)57),
-        PatchListWith26ControlPoints = unchecked((int)58),
-        PatchListWith27ControlPoints = unchecked((int)59),
-        PatchListWith28ControlPoints = unchecked((int)60),
-        PatchListWith29ControlPoints = unchecked((int)61),
-        PatchListWith30ControlPoints = unchecked((int)62),
-        PatchListWith31ControlPoints = unchecked((int)63),
-        PatchListWith32ControlPoints = unchecked((int)64)
+        Undefined = unchecked(0),
+        PointList = unchecked(1),
+        LineList = unchecked(2),
+        LineStrip = unchecked(3),
+        TriangleList = unchecked(4),
+        TriangleStrip = unchecked(5),
+        LineListAdjacency = unchecked(10),
+        LineStripAdjacency = unchecked(11),
+        TriangleListAdjacency = unchecked(12),
+        TriangleStripAdjacency = unchecked(13),
+        PatchListWith1ControlPoints = unchecked(33),
+        PatchListWith2ControlPoints = unchecked(34),
+        PatchListWith3ControlPoints = unchecked(35),
+        PatchListWith4ControlPoints = unchecked(36),
+        PatchListWith5ControlPoints = unchecked(37),
+        PatchListWith6ControlPoints = unchecked(38),
+        PatchListWith7ControlPoints = unchecked(39),
+        PatchListWith8ControlPoints = unchecked(40),
+        PatchListWith9ControlPoints = unchecked(41),
+        PatchListWith10ControlPoints = unchecked(42),
+        PatchListWith11ControlPoints = unchecked(43),
+        PatchListWith12ControlPoints = unchecked(44),
+        PatchListWith13ControlPoints = unchecked(45),
+        PatchListWith14ControlPoints = unchecked(46),
+        PatchListWith15ControlPoints = unchecked(47),
+        PatchListWith16ControlPoints = unchecked(48),
+        PatchListWith17ControlPoints = unchecked(49),
+        PatchListWith18ControlPoints = unchecked(50),
+        PatchListWith19ControlPoints = unchecked(51),
+        PatchListWith20ControlPoints = unchecked(52),
+        PatchListWith21ControlPoints = unchecked(53),
+        PatchListWith22ControlPoints = unchecked(54),
+        PatchListWith23ControlPoints = unchecked(55),
+        PatchListWith24ControlPoints = unchecked(56),
+        PatchListWith25ControlPoints = unchecked(57),
+        PatchListWith26ControlPoints = unchecked(58),
+        PatchListWith27ControlPoints = unchecked(59),
+        PatchListWith28ControlPoints = unchecked(60),
+        PatchListWith29ControlPoints = unchecked(61),
+        PatchListWith30ControlPoints = unchecked(62),
+        PatchListWith31ControlPoints = unchecked(63),
+        PatchListWith32ControlPoints = unchecked(64)
     }
 }
