@@ -1,19 +1,13 @@
 ﻿namespace HexaEngine.Editor
 {
-    using HexaEngine;
     using HexaEngine.Core;
     using HexaEngine.Core.Events;
     using HexaEngine.Core.Graphics;
     using HexaEngine.Core.Input;
-    using HexaEngine.D3D11;
     using HexaEngine.Rendering;
     using HexaEngine.Scenes;
     using System;
-    using System.Collections.Generic;
     using System.Diagnostics;
-    using System.Linq;
-    using System.Text;
-    using System.Threading.Tasks;
 
     public class EditorWindow : SdlWindow
     {
@@ -36,6 +30,12 @@
 
         public IGraphicsDevice Device => device ?? throw new InvalidOperationException();
         public IGraphicsContext Context => context ?? throw new InvalidOperationException();
+
+        public bool VSync;
+
+        public bool LimitFPS;
+
+        public int TargetFPS = 120;
 
         protected override void OnShown(ShownEventArgs args)
         {
@@ -95,7 +95,7 @@
                     }
 
                 renderer.EndDraw();
-                swapChain?.Present(Nucleus.Settings.VSync ? 1u : 0u);
+                swapChain?.Present(VSync ? 1u : 0u);
                 LimitFrameRate();
                 Keyboard.FrameUpdate();
             }
@@ -113,9 +113,9 @@
 
         private void LimitFrameRate()
         {
-            if (Nucleus.Settings.LimitFPS & !Nucleus.Settings.VSync)
+            if (LimitFPS & !VSync)
             {
-                int fps = Nucleus.Settings.TargetFPS;
+                int fps = TargetFPS;
                 long freq = Stopwatch.Frequency;
                 long frame = Stopwatch.GetTimestamp();
                 while ((frame - fpsStartTime) * fps < freq * fpsFrameCount)
