@@ -14,9 +14,9 @@ namespace ImGuiNET
         public ImGuiStoragePair* NativePtr { get; }
         public ImGuiStoragePairPtr(ImGuiStoragePair* nativePtr) => NativePtr = nativePtr;
         public ImGuiStoragePairPtr(IntPtr nativePtr) => NativePtr = (ImGuiStoragePair*)nativePtr;
-        public static implicit operator ImGuiStoragePairPtr(ImGuiStoragePair* nativePtr) => new ImGuiStoragePairPtr(nativePtr);
+        public static implicit operator ImGuiStoragePairPtr(ImGuiStoragePair* nativePtr) => new(nativePtr);
         public static implicit operator ImGuiStoragePair*(ImGuiStoragePairPtr wrappedPtr) => wrappedPtr.NativePtr;
-        public static implicit operator ImGuiStoragePairPtr(IntPtr nativePtr) => new ImGuiStoragePairPtr(nativePtr);
+        public static implicit operator ImGuiStoragePairPtr(IntPtr nativePtr) => new(nativePtr);
     }
 
     [StructLayout(LayoutKind.Explicit)]
