@@ -16,19 +16,19 @@ namespace ImPlotNET
         public ImPlotLimits* NativePtr { get; }
         public ImPlotLimitsPtr(ImPlotLimits* nativePtr) => NativePtr = nativePtr;
         public ImPlotLimitsPtr(IntPtr nativePtr) => NativePtr = (ImPlotLimits*)nativePtr;
-        public static implicit operator ImPlotLimitsPtr(ImPlotLimits* nativePtr) => new ImPlotLimitsPtr(nativePtr);
+        public static implicit operator ImPlotLimitsPtr(ImPlotLimits* nativePtr) => new(nativePtr);
         public static implicit operator ImPlotLimits* (ImPlotLimitsPtr wrappedPtr) => wrappedPtr.NativePtr;
-        public static implicit operator ImPlotLimitsPtr(IntPtr nativePtr) => new ImPlotLimitsPtr(nativePtr);
+        public static implicit operator ImPlotLimitsPtr(IntPtr nativePtr) => new(nativePtr);
         public ref ImPlotRange X => ref Unsafe.AsRef<ImPlotRange>(&NativePtr->X);
         public ref ImPlotRange Y => ref Unsafe.AsRef<ImPlotRange>(&NativePtr->Y);
         public bool Contains(ImPlotPoint p)
         {
-            byte ret = ImPlotNative.ImPlotLimits_ContainsPlotPoInt((ImPlotLimits*)(NativePtr), p);
+            byte ret = ImPlotNative.ImPlotLimits_ContainsPlotPoInt(NativePtr, p);
             return ret != 0;
         }
         public bool Contains(double x, double y)
         {
-            byte ret = ImPlotNative.ImPlotLimits_Containsdouble((ImPlotLimits*)(NativePtr), x, y);
+            byte ret = ImPlotNative.ImPlotLimits_Containsdouble(NativePtr, x, y);
             return ret != 0;
         }
     }
