@@ -16,14 +16,14 @@ namespace ImPlotNET
         public ImPlotPoint* NativePtr { get; }
         public ImPlotPointPtr(ImPlotPoint* nativePtr) => NativePtr = nativePtr;
         public ImPlotPointPtr(IntPtr nativePtr) => NativePtr = (ImPlotPoint*)nativePtr;
-        public static implicit operator ImPlotPointPtr(ImPlotPoint* nativePtr) => new ImPlotPointPtr(nativePtr);
+        public static implicit operator ImPlotPointPtr(ImPlotPoint* nativePtr) => new(nativePtr);
         public static implicit operator ImPlotPoint* (ImPlotPointPtr wrappedPtr) => wrappedPtr.NativePtr;
-        public static implicit operator ImPlotPointPtr(IntPtr nativePtr) => new ImPlotPointPtr(nativePtr);
+        public static implicit operator ImPlotPointPtr(IntPtr nativePtr) => new(nativePtr);
         public ref double x => ref Unsafe.AsRef<double>(&NativePtr->x);
         public ref double y => ref Unsafe.AsRef<double>(&NativePtr->y);
         public void Destroy()
         {
-            ImPlotNative.ImPlotPoint_destroy((ImPlotPoint*)(NativePtr));
+            ImPlotNative.ImPlotPoint_destroy(NativePtr);
         }
     }
 }
